api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "ansible-pylibssh>=1.3.0",
    "ansible-runner>=2.4.2",
    "fastapi[standard]>=0.124.4",
    "fastmcp>=2.14.0",
    "httpx>=0.28.1",
    "load-dotenv>=0.1.0",
    "openai>=2.13.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.45",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["ansible*", "configs*", "database*"]
